重新生成DO文件

+++ thrift-jooq/thrift/dao/struct/campaigndb/campaign_baidu_users_struct.thrift
@@ -6,7 +6,7 @@ struct CampaignBaiduUsersDO {
 
 	1: optional i32 id,	//null
 	2: optional i32 userId,	//null
-	3: optional i32 uid,	//baidu的uid
+	3: optional i64 uid,	//baidu的uid
 	4: optional string createTime	//null
 
 }
