+++ thrift-jooq/thrift/dao/struct/wordpressdb/wordpress_options_struct.thrift
@@ -4,7 +4,7 @@ namespace py thrift_gen.gen.dao.struct.wordpressdb
 
 struct WordpressOptionsDO {
 
-	1: optional i32 optionId,	//null
+	1: optional i64 optionId,	//null
 	2: optional string optionName,	//null
 	3: optional string optionValue,	//null
 	4: optional string autoload	//null
