+++ thrift-jooq/thrift/dao/struct/wordpressdb/wordpress_postmeta_struct.thrift
@@ -4,8 +4,8 @@ namespace py thrift_gen.gen.dao.struct.wordpressdb
 
 struct WordpressPostmetaDO {
 
-	1: optional i32 metaId,	//null
-	2: optional i32 postId,	//null
+	1: optional i64 metaId,	//null
+	2: optional i64 postId,	//null
 	3: optional string metaKey,	//null
 	4: optional string metaValue	//null
