+++ thrift-jooq/thrift/dao/struct/wordpressdb/wordpress_term_relationships_struct.thrift
@@ -4,8 +4,8 @@ namespace py thrift_gen.gen.dao.struct.wordpressdb
 
 struct WordpressTermRelationshipsDO {
 
-	1: optional i32 objectId,	//null
-	2: optional i32 termTaxonomyId,	//null
+	1: optional i64 objectId,	//null
+	2: optional i64 termTaxonomyId,	//null
 	3: optional i32 termOrder	//null
 
 }
