+++ thrift-jooq/thrift/dao/struct/wordpressdb/wordpress_termmeta_struct.thrift
@@ -4,8 +4,8 @@ namespace py thrift_gen.gen.dao.struct.wordpressdb
 
 struct WordpressTermmetaDO {
 
-	1: optional i32 metaId,	//null
-	2: optional i32 termId,	//null
+	1: optional i64 metaId,	//null
+	2: optional i64 termId,	//null
 	3: optional string metaKey,	//null
 	4: optional string metaValue	//null
