+++ thrift-jooq/thrift/dao/struct/wordpressdb/wordpress_terms_struct.thrift
@@ -4,9 +4,9 @@ namespace py thrift_gen.gen.dao.struct.wordpressdb
 
 struct WordpressTermsDO {
 
-	1: optional i32 termId,	//null
+	1: optional i64 termId,	//null
 	2: optional string name,	//null
 	3: optional string slug,	//null
-	4: optional i32 termGroup	//null
+	4: optional i64 termGroup	//null
 
 }
